+++ src/pages/User/Home.tsx
@@ -69,24 +69,22 @@ function Home() {
   const [isHover, setIsHover] = useState<boolean>(false);
 
   return (
-    <>
-      <AppContainer justifyValue={'center'} flexDirection={'column'} alignItems={'flex-start'}>
-        <>
-          <MainTitle>
-            키오스쿨로
-            <br /> 주점관리를 손쉽게,
-          </MainTitle>
-          <LinkAdminHome to={'/admin'} onMouseEnter={() => setIsHover(true)} onMouseLeave={() => setIsHover(false)}>
-            <LinkText>
-              내 주점 바로가기
-              <ArrowRight />
-            </LinkText>
-          </LinkAdminHome>
-          <AppFooter />
-        </>
-      </AppContainer>
-      <HoverOverlay isHover={isHover} />
-    </>
+    <AppContainer justifyValue={'center'} flexDirection={'column'} alignItems={'flex-start'}>
+      <>
+        <MainTitle>
+          키오스쿨로
+          <br /> 주점관리를 손쉽게,
+        </MainTitle>
+        <LinkAdminHome to={'/admin'} onMouseEnter={() => setIsHover(true)} onMouseLeave={() => setIsHover(false)}>
+          <LinkText>
+            내 주점 바로가기
+            <ArrowRight />
+          </LinkText>
+        </LinkAdminHome>
+        <AppFooter />
+        <HoverOverlay isHover={isHover} />
+      </>
+    </AppContainer>
   );
 }
